Adjust scalabel.label functions' locations
